requirements.txt
Django>=5.1,<6.0
gunicorn>=22.0.0,<21.0
psycopg2-binary<3.0,>=2.9
python-dotenv>=0.19.2,<1.0
twilio>=8.0.0,<9.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt~=5.3.1